client: Mac build fixes
compareOSVersion() is a utility function.  Put it in an appropriate file.

Declare compareOSVersion() in the .h file for its .cpp file.
Do not use "extern" in a .cpp file.
