+++ client/cs_platforms.cpp
@@ -42,11 +42,11 @@ LPFN_ISWOW64PROCESS fnIsWow64Process;
 
 #ifdef __APPLE__
 #include <sys/sysctl.h>
-extern int compareOSVersionTo(int toMajor, int toMinor);
 #endif
 
 #include "error_numbers.h"
 #include "filesys.h"
+#include "hostinfo.h"
 #include "str_util.h"
 #include "str_replace.h"
 #include "util.h"
